Split dream list pagination helpers
